validate on controller, wrap form groups and display error

+++ BPOR/BPOR.Rms/ParticipantQueryExtensions.cs
@@ -25,7 +25,7 @@ public static class ParticipantQueryExtensions
         if (selectedVolunteersPreferredContact.HasValue)
         {
 
-            return query.Where(x => x.ContactMethods.Any(e => e.ContactMethodId == (int)selectedVolunteersPreferredContact));
+            return query.Where(x => x.ContactMethods.Any(e => e.ContactMethodId == selectedVolunteersPreferredContact));
         }
         return query;
     }
